+++ contracts/src/MarketFactory.sol
@@ -9,10 +9,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.20;
 
-import "@openzeppelin/contracts/proxy/Clones.sol";
+import {IConditionalTokens, IRealityETH_v3_0, IWrapped1155Factory} from "./Interfaces.sol";
 import "./Market.sol";
 import "./RealityProxy.sol";
-import {IRealityETH_v3_0, IConditionalTokens, IWrapped1155Factory} from "./Interfaces.sol";
+import "@openzeppelin/contracts/proxy/Clones.sol";
 
 contract MarketFactory {
     using Clones for address;
@@ -100,143 +100,90 @@ contract MarketFactory {
 
     /// @dev Creates a Categorical market.
     /// @notice Categorical markets are associated with a Reality question that has only one answer.
-    function createCategoricalMarket(
-        CreateMarketParams calldata params
-    ) external returns (address) {
-        require(
-            params.outcomes.length >= 2,
-            "Outcomes count must be 2 or more"
-        );
+    function createCategoricalMarket(CreateMarketParams calldata params) external returns (address) {
+        require(params.outcomes.length >= 2, "Outcomes count must be 2 or more");
 
         string[] memory encodedQuestions = new string[](1);
-        encodedQuestions[0] = encodeRealityQuestionWithOutcomes(
-            params.marketName,
-            params.outcomes,
-            params.category,
-            params.lang
-        );
+        encodedQuestions[0] =
+            encodeRealityQuestionWithOutcomes(params.marketName, params.outcomes, params.category, params.lang);
 
-        return
-            createMarket(
-                params,
-                params.marketName,
-                InternalMarketConfig({
-                    encodedQuestions: encodedQuestions,
-                    outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
-                    templateId: REALITY_SINGLE_SELECT_TEMPLATE
-                })
-            );
+        return createMarket(
+            params,
+            params.marketName,
+            InternalMarketConfig({
+                encodedQuestions: encodedQuestions,
+                outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
+                templateId: REALITY_SINGLE_SELECT_TEMPLATE
+            })
+        );
     }
 
     /// @dev Creates a Multi Categorical market.
     /// @notice Multi Categorical markets are associated with a Reality question that has one or more answers.
-    function createMultiCategoricalMarket(
-        CreateMarketParams calldata params
-    ) external returns (address) {
-        require(
-            params.outcomes.length >= 2,
-            "Outcomes count must be 2 or more"
-        );
+    function createMultiCategoricalMarket(CreateMarketParams calldata params) external returns (address) {
+        require(params.outcomes.length >= 2, "Outcomes count must be 2 or more");
 
         string[] memory encodedQuestions = new string[](1);
-        encodedQuestions[0] = encodeRealityQuestionWithOutcomes(
-            params.marketName,
-            params.outcomes,
-            params.category,
-            params.lang
-        );
+        encodedQuestions[0] =
+            encodeRealityQuestionWithOutcomes(params.marketName, params.outcomes, params.category, params.lang);
 
-        return
-            createMarket(
-                params,
-                params.marketName,
-                InternalMarketConfig({
-                    encodedQuestions: encodedQuestions,
-                    outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
-                    templateId: REALITY_MULTI_SELECT_TEMPLATE
-                })
-            );
+        return createMarket(
+            params,
+            params.marketName,
+            InternalMarketConfig({
+                encodedQuestions: encodedQuestions,
+                outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
+                templateId: REALITY_MULTI_SELECT_TEMPLATE
+            })
+        );
     }
 
     /// @dev Creates a Scalar market.
     /// @notice Scalar markets are associated with a Reality question that resolves to a numeric value.
-    function createScalarMarket(
-        CreateMarketParams calldata params
-    ) external returns (address) {
-        require(
-            params.upperBound > params.lowerBound,
-            "upperBound must be higher than lowerBound"
-        );
+    function createScalarMarket(CreateMarketParams calldata params) external returns (address) {
+        require(params.upperBound > params.lowerBound, "upperBound must be higher than lowerBound");
         // values reserved by Reality for INVALID and UNRESOLVED_ANSWER.
-        require(
-            params.upperBound < type(uint256).max - 2,
-            "upperBound must be less than uint256.max - 2"
-        );
+        require(params.upperBound < type(uint256).max - 2, "upperBound must be less than uint256.max - 2");
         require(params.outcomes.length == 2, "Outcomes count must be 2");
 
         string[] memory encodedQuestions = new string[](1);
-        encodedQuestions[0] = encodeRealityQuestionWithoutOutcomes(
-            params.marketName,
-            params.category,
-            params.lang
-        );
+        encodedQuestions[0] = encodeRealityQuestionWithoutOutcomes(params.marketName, params.category, params.lang);
 
-        return
-            createMarket(
-                params,
-                params.marketName,
-                InternalMarketConfig({
-                    encodedQuestions: encodedQuestions,
-                    outcomeSlotCount: 3, // additional outcome for Invalid Result.
-                    templateId: REALITY_UINT_TEMPLATE
-                })
-            );
+        return createMarket(
+            params,
+            params.marketName,
+            InternalMarketConfig({
+                encodedQuestions: encodedQuestions,
+                outcomeSlotCount: 3, // additional outcome for Invalid Result.
+                templateId: REALITY_UINT_TEMPLATE
+            })
+        );
     }
 
     /// @dev Creates a Multi Scalar market.
     /// @notice Multi Scalar markets are associated with two or more Reality questions, and each one of them resolves to a numeric value.
-    function createMultiScalarMarket(
-        CreateMarketParams calldata params
-    ) external returns (address) {
-        require(
-            params.outcomes.length >= 2,
-            "Outcomes count must be 2 or more"
-        );
+    function createMultiScalarMarket(CreateMarketParams calldata params) external returns (address) {
+        require(params.outcomes.length >= 2, "Outcomes count must be 2 or more");
 
         string[] memory encodedQuestions = new string[](params.outcomes.length);
 
         for (uint256 i = 0; i < params.outcomes.length; i++) {
             encodedQuestions[i] = encodeRealityQuestionWithoutOutcomes(
-                string(
-                    abi.encodePacked(
-                        params.questionStart,
-                        params.outcomes[i],
-                        params.questionEnd
-                    )
-                ),
+                string(abi.encodePacked(params.questionStart, params.outcomes[i], params.questionEnd)),
                 params.category,
                 params.lang
             );
         }
 
-        return
-            createMarket(
-                params,
-                string(
-                    abi.encodePacked(
-                        params.questionStart,
-                        "[",
-                        params.outcomeType,
-                        "]",
-                        params.questionEnd
-                    )
-                ),
-                InternalMarketConfig({
-                    encodedQuestions: encodedQuestions,
-                    outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
-                    templateId: REALITY_UINT_TEMPLATE
-                })
-            );
+        return createMarket(
+            params,
+            string(abi.encodePacked(params.questionStart, "[", params.outcomeType, "]", params.questionEnd)),
+            InternalMarketConfig({
+                encodedQuestions: encodedQuestions,
+                outcomeSlotCount: params.outcomes.length + 1, // additional outcome for Invalid Result.
+                templateId: REALITY_UINT_TEMPLATE
+            })
+        );
     }
 
     /// @dev Creates the Market and deploys the wrapped ERC20 tokens.
@@ -245,10 +192,8 @@ contract MarketFactory {
         string memory marketName,
         InternalMarketConfig memory config
     ) internal returns (address) {
-        (
-            Market.ConditionalTokensParams memory conditionalTokensParams,
-            Market.RealityParams memory realityParams
-        ) = createNewMarketParams(params, config);
+        (Market.ConditionalTokensParams memory conditionalTokensParams, Market.RealityParams memory realityParams) =
+            createNewMarketParams(params, config);
 
         Market instance = Market(market.clone());
 
@@ -280,13 +225,7 @@ contract MarketFactory {
     function createNewMarketParams(
         CreateMarketParams memory params,
         InternalMarketConfig memory config
-    )
-        internal
-        returns (
-            Market.ConditionalTokensParams memory,
-            Market.RealityParams memory
-        )
-    {
+    ) internal returns (Market.ConditionalTokensParams memory, Market.RealityParams memory) {
         bytes32 parentCollectionId = params.parentMarket == address(0)
             ? bytes32(0)
             : conditionalTokens.getCollectionId(
@@ -295,40 +234,21 @@ contract MarketFactory {
                 1 << params.parentOutcome
             );
 
-        bytes32[] memory questionsIds = new bytes32[](
-            config.encodedQuestions.length
-        );
+        bytes32[] memory questionsIds = new bytes32[](config.encodedQuestions.length);
 
         for (uint256 i = 0; i < config.encodedQuestions.length; i++) {
-            questionsIds[i] = askRealityQuestion(
-                config.encodedQuestions[i],
-                config.templateId,
-                params.openingTime,
-                params.minBond
-            );
+            questionsIds[i] =
+                askRealityQuestion(config.encodedQuestions[i], config.templateId, params.openingTime, params.minBond);
         }
 
         // questionId must be a hash of all the values that RealityProxy.resolve() uses to resolve a market, this way if an attacker tries to resolve a fake market by changing some value its questionId will not match the id of a valid market.
         bytes32 questionId = keccak256(
-            abi.encode(
-                questionsIds,
-                params.outcomes.length,
-                config.templateId,
-                params.lowerBound,
-                params.upperBound
-            )
+            abi.encode(questionsIds, params.outcomes.length, config.templateId, params.lowerBound, params.upperBound)
         );
-        bytes32 conditionId = prepareCondition(
-            questionId,
-            config.outcomeSlotCount
-        );
+        bytes32 conditionId = prepareCondition(questionId, config.outcomeSlotCount);
 
-        (IERC20[] memory wrapped1155, bytes[] memory data) = deployERC20Positions(
-            parentCollectionId,
-            conditionId,
-            config.outcomeSlotCount,
-            params.tokenNames
-        );
+        (IERC20[] memory wrapped1155, bytes[] memory data) =
+            deployERC20Positions(parentCollectionId, conditionId, config.outcomeSlotCount, params.tokenNames);
 
         return (
             Market.ConditionalTokensParams({
@@ -366,26 +286,10 @@ contract MarketFactory {
         bytes memory encodedOutcomes = abi.encodePacked('"', outcomes[0], '"');
 
         for (uint256 i = 1; i < outcomes.length; i++) {
-            encodedOutcomes = abi.encodePacked(
-                encodedOutcomes,
-                ',"',
-                outcomes[i],
-                '"'
-            );
+            encodedOutcomes = abi.encodePacked(encodedOutcomes, ',"', outcomes[i], '"');
         }
 
-        return
-            string(
-                abi.encodePacked(
-                    question,
-                    separator,
-                    encodedOutcomes,
-                    separator,
-                    category,
-                    separator,
-                    lang
-                )
-            );
+        return string(abi.encodePacked(question, separator, encodedOutcomes, separator, category, separator, lang));
     }
 
     /// @dev Encodes the question, category and language following the Reality structure.
@@ -401,10 +305,7 @@ contract MarketFactory {
     ) internal pure returns (string memory) {
         bytes memory separator = abi.encodePacked(unicode"\u241f");
 
-        return
-            string(
-                abi.encodePacked(question, separator, category, separator, lang)
-            );
+        return string(abi.encodePacked(question, separator, category, separator, lang));
     }
 
     /// @dev Asks a question on reality.
@@ -419,19 +320,11 @@ contract MarketFactory {
         uint32 openingTime,
         uint256 minBond
     ) internal returns (bytes32) {
-        bytes32 content_hash = keccak256(
-            abi.encodePacked(templateId, openingTime, encodedQuestion)
-        );
+        bytes32 content_hash = keccak256(abi.encodePacked(templateId, openingTime, encodedQuestion));
 
         bytes32 question_id = keccak256(
             abi.encodePacked(
-                content_hash,
-                arbitrator,
-                questionTimeout,
-                minBond,
-                address(realitio),
-                address(this),
-                uint256(0)
+                content_hash, arbitrator, questionTimeout, minBond, address(realitio), address(this), uint256(0)
             )
         );
 
@@ -439,37 +332,19 @@ contract MarketFactory {
             return question_id;
         }
 
-        return
-            realitio.askQuestionWithMinBond(
-                templateId,
-                encodedQuestion,
-                arbitrator,
-                questionTimeout,
-                openingTime,
-                0,
-                minBond
-            );
+        return realitio.askQuestionWithMinBond(
+            templateId, encodedQuestion, arbitrator, questionTimeout, openingTime, 0, minBond
+        );
     }
 
     /// @dev Prepares the CTF condition and returns the conditionId.
     /// @param questionId An identifier for the question to be answered by the oracle.
     /// @param outcomeSlotCount The number of outcome slots which must be used for this condition. Must not exceed 256.
-    function prepareCondition(
-        bytes32 questionId,
-        uint outcomeSlotCount
-    ) internal returns (bytes32) {
-        bytes32 conditionId = conditionalTokens.getConditionId(
-            address(realityProxy),
-            questionId,
-            outcomeSlotCount
-        );
+    function prepareCondition(bytes32 questionId, uint256 outcomeSlotCount) internal returns (bytes32) {
+        bytes32 conditionId = conditionalTokens.getConditionId(address(realityProxy), questionId, outcomeSlotCount);
 
         if (conditionalTokens.getOutcomeSlotCount(conditionId) == 0) {
-            conditionalTokens.prepareCondition(
-                address(realityProxy),
-                questionId,
-                outcomeSlotCount
-            );
+            conditionalTokens.prepareCondition(address(realityProxy), questionId, outcomeSlotCount);
         }
 
         return conditionId;
@@ -491,21 +366,11 @@ contract MarketFactory {
         wrapped1155 = new IERC20[](outcomeSlotCount);
         data = new bytes[](outcomeSlotCount);
 
-        for (uint j = 0; j < outcomeSlotCount; j++) {
-            bytes32 collectionId = conditionalTokens.getCollectionId(
-                parentCollectionId,
-                conditionId,
-                1 << j
-            );
-            uint256 tokenId = conditionalTokens.getPositionId(
-                collateralToken,
-                collectionId
-            );
+        for (uint256 j = 0; j < outcomeSlotCount; j++) {
+            bytes32 collectionId = conditionalTokens.getCollectionId(parentCollectionId, conditionId, 1 << j);
+            uint256 tokenId = conditionalTokens.getPositionId(collateralToken, collectionId);
 
-            require(
-                j == invalidResultIndex || bytes(tokenNames[j]).length != 0,
-                "Missing token name"
-            );
+            require(j == invalidResultIndex || bytes(tokenNames[j]).length != 0, "Missing token name");
 
             bytes memory _data = abi.encodePacked(
                 toString31(j == invalidResultIndex ? "SEER_INVALID_RESULT" : tokenNames[j]),
@@ -513,11 +378,7 @@ contract MarketFactory {
                 uint8(18)
             );
 
-            IERC20 _wrapped1155 = wrapped1155Factory.requireWrapped1155(
-                address(conditionalTokens),
-                tokenId,
-                _data
-            );
+            IERC20 _wrapped1155 = wrapped1155Factory.requireWrapped1155(address(conditionalTokens), tokenId, _data);
 
             wrapped1155[j] = _wrapped1155;
             data[j] = _data;
@@ -527,9 +388,7 @@ contract MarketFactory {
     /// @dev Encodes a short string (less than than 31 bytes long) as for storage as expected by Solidity.
     /// See https://github.com/gnosis/1155-to-20/pull/4#discussion_r573630922
     /// @param value String to encode.
-    function toString31(
-        string memory value
-    ) internal pure returns (bytes32 encodedString) {
+    function toString31(string memory value) internal pure returns (bytes32 encodedString) {
         uint256 length = bytes(value).length;
         require(length < 32, "string too long");
